test(vitest): port activeRuntimeNotebookContextManager to Vitest
